[API] consolidate types into use sites
