+++ apps/web/src/components/maps/Map.tsx
@@ -3,7 +3,7 @@ import type { MapRef } from 'react-map-gl';
 import { AttributionControl, Map as MapGL } from 'react-map-gl';
 import { Box, useTheme } from '@mui/material';
 import { useColorScheme } from 'ui/theme/useColorScheme';
-import type { MapLocation } from 'api/types/MapLocation';
+import type { MapLocation } from 'api/server/contentful/MapLocation';
 import { StandardControls } from './StandardControls';
 
 export type MapProps = {
