+++ apps/web/src/components/spotify/PlaybackStatus.tsx
@@ -1,7 +1,7 @@
 import { Music } from 'lucide-react';
 import { Typography } from '@mui/material';
 import { HorizontalStack } from 'ui/core/HorizontalStack';
-import type { Track } from 'api/types/spotify/Track';
+import type { Track } from 'api/server/spotify/Track';
 import { useRelativeTimeFormat } from 'hooks/useRelativeTimeFormat';
 
 type PlaybackStatusProps = {
