+++ apps/web/src/components/spotify/TrackListing.tsx
@@ -1,6 +1,6 @@
 import { Stack } from '@mui/material';
 import { HorizontalStack } from 'ui/core/HorizontalStack';
-import type { Track } from 'api/types/spotify/Track';
+import type { Track } from 'api/server/spotify/Track';
 import { AlbumImage } from './AlbumImage';
 import { ArtistList } from './ArtistList';
 import { SpotifyLogo } from './SpotifyLogo';
