+++ apps/web/src/pages/api/webhooks.ts
@@ -3,7 +3,7 @@ import { isRecord } from 'shared-core/helpers/typeguards';
 import { handleApiError, methodNotAllowedError } from 'api/handleApiError';
 import { echoStravaChallengeIfValid } from 'api/server/strava/echoStravaChallengeIfValid';
 import { syncStravaWebhookUpdateWithDb } from 'api/server/strava/syncStravaWebhookUpdateWithDb';
-import type { StravaWebhookEvent } from 'api/types/StravaWebhookEvent';
+import type { StravaWebhookEvent } from 'api/server/strava/StravaWebhookEvent';
 
 // Just a shorthand for this function type
 type Processor = (request: NextApiRequest, response: NextApiResponse) => void;
