+++ packages/tsconfig/nextjs.json
@@ -17,6 +17,6 @@
       }
     ]
   },
-  "include": ["src", "next-env.d.ts"],
+  "include": ["src", "next.config.js", "next-env.d.ts"],
   "exclude": ["node_modules", "**/*/dist", "**/*/build", "**/*/.next"]
 }
